test(web-terminal): wait on the dock's own settled signal before a visual capture
